Refactored to separate project for github support
