+++ scalafix/readme.md
@@ -1,0 +1,7 @@
+# Scalafix rules for http4s
+
+To develop rule:
+```
+sbt ~tests/test
+# edit rules/src/main/scala/fix/Http4s.scala
+```
