--- scalafix/src/main/resources/META-INF/services/scalafix.v1.Rule
@@ -1,1 +1,0 @@
-fix.Http4s018To020
