+++ scalafix/tests/src/test/scala/fix/RuleSuite.scala
@@ -1,0 +1,7 @@
+package fix
+
+import scalafix.testkit.SemanticRuleSuite
+
+class RuleSuite extends SemanticRuleSuite() {
+  runAllTests()
+}
